Glaive work

+++ Common/Projectiles/ProjectileType/GlaiveProjectile.cs
@@ -5,11 +5,16 @@ using Terraria.ModLoader;
 using Terraria.Audio;
 using Terraria.Enums;
 using Microsoft.Xna.Framework;
+using DestinyMod.Content.Projectiles.Weapons.Melee.Glaive;
 
 namespace DestinyMod.Common.Projectiles.ProjectileType
 {
 	public abstract class GlaiveProjectile : DestinyModProjectile
 	{
+		protected int SpriteWidth;
+
+		protected int SpriteHeight;
+
 		public override void AutomaticSetDefaults()
 		{
 			Projectile.friendly = true;
@@ -20,26 +25,35 @@ namespace DestinyMod.Common.Projectiles.ProjectileType
 			Projectile.DamageType = DamageClass.Melee;
 			Projectile.extraUpdates = 1;
 			Projectile.timeLeft = 360;
-			//Projectile.aiStyle = -1;
+			Projectile.aiStyle = -1;
 		}
 
         public override void AI()
         {
 			Player owner = Main.player[Projectile.owner];
-			Projectile.rotation = Projectile.velocity.ToRotation() + MathHelper.PiOver2;
-			Projectile.ai[0]++;
-			Projectile.Opacity = Utils.GetLerpValue(0, 7, Projectile.ai[0], true) * Utils.GetLerpValue(16, 12, Projectile.ai[0], true);
-			Projectile.Center = owner.RotatedRelativePoint(owner.MountedCenter, false, false) + Projectile.velocity * (Projectile.ai[0] - 1);
-			Projectile.spriteDirection = Vector2.Dot(Projectile.velocity, Vector2.UnitX) >= 0f ? 1 : -1;
-			if (Projectile.ai[0] >= 16)
-            {
+
+			if (++Projectile.ai[0] >= 16 && Projectile.ai[1] != 2) //important
+			{
+				Projectile.NewProjectile(owner.GetProjectileSource_Misc(0), owner.Center, Projectile.velocity * 3, ModContent.ProjectileType<GlaiveShot>(), 10, 0, owner.whoAmI);
 				Projectile.Kill();
-            }
+			}
 			else
-            {
+			{
 				owner.heldProj = Projectile.whoAmI;
-            }
+				owner.itemTime = owner.itemAnimation = 1;
+			}
+
+			Projectile.Opacity = Utils.GetLerpValue(0, 7, Projectile.ai[0], true) * Utils.GetLerpValue(16, 12, Projectile.ai[0], true);
+			Projectile.rotation = Projectile.velocity.ToRotation() + MathHelper.PiOver2 - MathHelper.PiOver4 * Projectile.spriteDirection;
+			Projectile.Center = owner.RotatedRelativePoint(owner.MountedCenter, false, false) + Projectile.velocity * (Projectile.ai[0] - 1);
+			Projectile.spriteDirection = (Vector2.Dot(Projectile.velocity, Vector2.UnitX) >= 0f).ToDirectionInt();
+
+			DrawOriginOffsetX = 0;
+			DrawOffsetX = -(SpriteWidth / 2 - Projectile.width / 2);
+			DrawOriginOffsetY = -(SpriteHeight / 2 - Projectile.height / 2);
         }
+
+		public override bool ShouldUpdatePosition() => false;
 
         public override bool? Colliding(Rectangle projHitbox, Rectangle targetHitbox)
         {
